'<div?>' fix and removing redundant validation
I found a bug in this package that causes atom to freeze when the tag
entered is invalid, e.g. if you accidentally type '<div?', when you hit
'>', atom will freeze up. I've fixed this with a try catch, it now will
always closes this kind of malformed tag on a new line, as I assume
malformed tags will not be in the forceBlock or forceInline objects.

While I was fixing this, I noted that e?.newText is tested for '>' twice
(once before execAutoclose, and once at the beginning of execAutoclose),
so I removed the check before calling execAutoclose. It will now always
call execAutoclose on new character, but immediately return void if the
character is not '>'.

This is my first pull request, so if I've broken any etiquette, please
do correct me.